Cleanup
Reorder files to keep same folder structure as assets
Cleanup comments and unnecessary code

[ADD] Create compatible geometry function in PieceMaterial
