+++ rvg-portfolio/src/Assets/Models/Chess/Pieces/Cup.glb
