+++ rvg-portfolio/src/components/Page/Overlay.jsx
@@ -16,6 +16,7 @@ const Overlay = forwardRef(({ caption, scroll }, ref) => (
     >
         {/* TODO DELETE */}
         <span className='fixed text-white' ref={caption}>0.00</span>
+        
         <Home className='relative'/>
         <Projects className='relative'/>
         <AboutMe className='relative'/>
