+++ rvg-portfolio/src/components/Page/Projects.jsx
@@ -3,8 +3,6 @@ import React from "react";
 import classnames from 'classnames'
 
 //overscroll-auto
-
-
 function Projects({className}) {
     return(
         <section className={classnames(className, 'w-screen h-screen justify-center')}>
