+++ rvg-portfolio/src/components/Page/Scene.jsx
@@ -1,19 +1,16 @@
 import React, { Suspense } from 'react'
 import { Canvas } from '@react-three/fiber'
-import { OrbitControls, OrthographicCamera } from '@react-three/drei'
 import classnames from 'classnames'
 
+//Temp
+import Fox from '../Models/Fox'
+import ChessboardWithMaterial from '../Models/Chess/ChessboardWithMaterial'
+
 //Models
-import Chessboard from '../Models/Chessboard'
-import ChessboardWithMaterial from '../Models/ChessboardWithMaterial'
-import Box from '../Models/Box'
-import Cup from '../Models/Pieces/Cup'
-
-
+import Chessboard from '../Models/Chess/Chessboard'
+import Cup from '../Models/Chess/Pieces/Cup'
 
 function Scene({ overlay, className }) {
-    
-    // console.log(overlay)
     return (
         <div className={classnames(className, 'w-full h-screen')}>
             {/* anti-alliasing is set to true by default */}
@@ -24,17 +21,14 @@ function Scene({ overlay, className }) {
                 }}
                 raycaster={{ computeOffsets: ({ clientX, clientY }) => ({ offsetX: clientX, offsetY: clientY }) }}
             > 
-                {/* <OrthographicCamera makeDefault zoom={15} position={[0, 45, 35]}/> */}
-                {/* <OrbitControls enablePan={true} enableRotate={true} /> */}
                 <gridHelper />
                 <ambientLight />
-                {/* <pointLight position={[10, 10, 10]} /> */}
+                <pointLight position={[10, 10, 10]} />
                 <Suspense fallback={null}>
                     <Chessboard position={[0, 0, 35]} rotation={[0, Math.PI/4, 0]} />
                     <Cup/>
-                    {/* <Box/> */}
+                    <Fox/>
                 </Suspense>
-                {/* <ChessboardWithMaterial position={[0, 0, 30]} rotation={[0, Math.PI/4, 0]} /> */}
             </Canvas>
         </div>
     )
